src/components/ca/details/IssuedCertificatesTab.tsx and src/app/certific
